Synchronize code for OPPO CPH2695_16.0.2.400(SP01EX01) PKV110_16.0.2.400(SP01CN01) Based on MTK release TAG:t-alps-release-t0.mp1.tc16sp-pr5-V1.62

+++ drivers/misc/mediatek/imgsensor/src-v4l2/common/imx888_mipi_raw/imx888mipiraw_Sensor.c
@@ -967,7 +967,7 @@ static void imx888_seamless_switch(struct subdrv_ctx *ctx, u8 *para, u32 *len)
 			u32 frame_length_in_lut[IMGSENSOR_STAGGER_EXPOSURE_CNT] = {0};
 
 			set_multi_shutter_frame_length_in_lut(ctx,
-			ae_ctrl, 2, 0, frame_length_in_lut);
+			(u64)ae_ctrl, 2, 0, frame_length_in_lut);
 			set_multi_gain_in_lut(ctx, ae_ctrl + 5, 2);
 		}
 			break;
@@ -976,7 +976,7 @@ static void imx888_seamless_switch(struct subdrv_ctx *ctx, u8 *para, u32 *len)
 			u32 frame_length_in_lut[IMGSENSOR_STAGGER_EXPOSURE_CNT] = {0};
 
 			set_multi_shutter_frame_length_in_lut(ctx,
-			ae_ctrl, 3, 0, frame_length_in_lut);
+			(u64)ae_ctrl, 3, 0, frame_length_in_lut);
 			set_multi_gain_in_lut(ctx, ae_ctrl + 5, 3);
 		}
 			break;
